plainbox: move Colorizer from impl.commands.inv_run to impl.color
